# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LinkedIn GPT API",
    "description": "API for accessing LinkedIn data, including top voices, search, profiles, and authentication.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "https://db828f2e-5d46-40d2-9a43-3f782fc3f1d8-00-2nhlwjy3wyleq.sisko.replit.dev/api/v1"
    }
  ],
  "paths": {
...
    "/api/v1/license/verify": {
      "post": {
        "description": "Verify a Gumroad subscription using email, enables premium features",
        "operationId": "verifySubscription",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "User's email address to verify subscription"
                  }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "subscriptionId": {
                      "type": "string",
                      "description": "Subscription ID to use for premium features"
                    },
                    "subscriptionType": {
                      "type": "string",
                      "enum": ["free", "premium"]
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or no active subscription found"
          },
          "500": {
            "description": "Server error during subscription verification"
          }
        },
        "tags": ["License"]
      }
    },
    "/api/v1/license/check/{email}": {
      "get": {
        "description": "Check if a user has a valid subscription",
        "operationId": "checkSubscription",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "User's email to check subscription for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Valid subscription found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "subscriptionId": {
                      "type": "string"
                    },
                    "subscriptionType": {
                      "type": "string",
                      "enum": ["free", "premium"]
                    },
                    "verifiedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No valid subscription found for this user"
          },
          "500": {
            "description": "Server error during subscription check"
          }
        },
        "tags": ["License"]
      }
    }
  },
...
}